Add travis-ci configuration file (sodaq boards don't work yet) and fix a compile error in one example

+++ examples/TheThingsUno-GPSshield-TTN-Mapper-binary/TheThingsUno-GPSshield-TTN-Mapper-binary.ino
@@ -164,7 +164,7 @@ void build_packet()
   txBuffer[8] = hdopGps & 0xFF;
 
   toLog = "";
-  for(int i = 0; i<sizeof(txBuffer); i++)
+  for(size_t i = 0; i<sizeof(txBuffer); i++)
   {
     char buffer[3];
     sprintf(buffer, "%02x", txBuffer[i]);
